# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Employee Management System API
  description: |
    RESTful API for managing employees, users, leave requests, and authentication.
    
    Features:
    - Employee CRUD operations with CSV import/export
    - User management with role-based access control
    - Leave request workflow with approval/rejection
    - Email verification and password reset
    - Rate limiting and caching
  version: 1.0.0
  contact:
    name: EMS Development Team
  license:
    name: MIT License
servers:
  - url: https://employee-management-system-production-2a1a.up.railway.app/api/v1
    description: Production server (Railway)
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - cookieAuth: []
...
paths:
...
  /employees:
    get:
      tags:
        - Employee Management
      summary: Get all employees
      description: Retrieve paginated list of employees with optional filtering
      operationId: getAllEmployees
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
            maximum: 20
        - name: sortBy
          in: query
          schema:
            type: string
            default: id
        - name: firstName
          in: query
          schema:
            type: string
        - name: lastName
          in: query
          schema:
            type: string
        - name: email
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: minSalary
          in: query
          schema:
            type: number
            format: float
        - name: maxSalary
          in: query
          schema:
            type: number
            format: float
      responses:
        '200':
          description: Employees retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedEmployeeDto'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Employee Management
      summary: Create employee
      description: Create a new employee with user credentials (ADMIN only)
      operationId: createEmployee
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeWithUserCreationDto'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
          headers:
            Location:
              schema:
                type: string
              description: URL of the newly created employee
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /employees/{id}:
    get:
      tags:
        - Employee Management
      summary: Get employee by ID
      operationId: getEmployeeById
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Employee retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Employee Management
      summary: Update employee (full)
      operationId: updateEmployee
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeDto'
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'

    patch:
      tags:
        - Employee Management
      summary: Update employee (partial)
      operationId: patchUpdateEmployee
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'

    delete:
      tags:
        - Employee Management
      summary: Delete employee
      operationId: deleteEmployee
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Employee deleted successfully
  /employees/upload:
    post:
      tags:
        - Employee Management
      summary: Upload employees from CSV
      operationId: uploadEmployees
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Employees uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  count:
                    type: integer
                  employees:
                    type: array
                    items:
                      $ref: '#/components/schemas/Employee'
  /employees/export:
    get:
      tags:
        - Employee Management
      summary: Export employees to CSV
      operationId: exportEmployees
      security:
        - cookieAuth: []
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary

  # User Management
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Session-based authentication
  schemas:
...
    EmployeeDto:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        phoneNumber:
          type: string
          example: "1234567890"
        hireDate:
          type: string
          format: date
          example: "2023-01-15"
        jobTitle:
          type: string
          example: "Software Engineer"
        salary:
          type: number
          format: float
          example: 75000.00
        department:
          type: string
          example: "IT"
    EmployeeWithUserCreationDto:
      type: object
      properties:
        employee:
          $ref: '#/components/schemas/EmployeeDto'
        username:
          type: string
        temporaryPassword:
          type: string
        roles:
          type: array
          items:
            type: string
        authorities:
          type: array
          items:
            type: string
    Employee:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
    PagedEmployeeDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeDto'
        totalElements:
          type: integer
        totalPages:
          type: integer
        pageable:
          type: object
          properties:
            pageNumber:
              type: integer
            pageSize:
              type: integer
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        validationErrors:
          type: object
          additionalProperties:
            type: string
...